Move prepare function to a dedicated go file
